+++ app/gui2/src/components/visualizations/LoadingErrorVisualization.vue
@@ -4,7 +4,6 @@ export const inputType = 'Any'
 </script>
 
 <script setup lang="ts">
-import { VisualizationContainer } from '@/util/visualizationBuiltins'
 import { watchEffect } from 'vue'
 
 const props = defineProps<{ data: { name: string; error: Error } }>()
@@ -13,14 +12,12 @@ watchEffect(() => console.error(props.data.error))
 </script>
 
 <template>
-  <VisualizationContainer :belowToolbar="true">
-    <div class="LoadingErrorVisualization">
-      <div>
-        <span>Could not load visualization '<span v-text="props.data.name"></span>':</span>
-        <span v-text="props.data.error.message"></span>
-      </div>
+  <div class="LoadingErrorVisualization">
+    <div>
+      <span>Could not load visualization '<span v-text="props.data.name"></span>':</span>
+      <span v-text="props.data.error.message"></span>
     </div>
-  </VisualizationContainer>
+  </div>
 </template>
 
 <style scoped>
