+++ app/gui2/src/util/visualizationBuiltins.ts
@@ -1,3 +1,2 @@
-export { default as VisualizationContainer } from '@/components/VisualizationContainer.vue'
 export { useVisualizationConfig } from '@/providers/visualizationConfig'
 export { defineKeybinds } from '@/util/shortcuts'
