+++ app/gui2/stories/story.css
@@ -20,7 +20,7 @@
   cursor: pointer !important;
 }
 
-.VisualizationContainer {
+.GraphVisualization {
   z-index: 0 !important;
   min-width: 0 !important;
 }
