+++ app/gui2/templates/visualization.vue
@@ -16,7 +16,6 @@ interface Data {
 <script setup lang="ts">
 import { onMounted } from 'vue'
 
-import { VisualizationContainer } from 'builtins'
 // Optional: add your own external dependencies.
 // import dependency from 'https://<js dependency here>'
 //
@@ -43,10 +42,8 @@ onMounted(() => {
 </script>
 
 <template>
-  <VisualizationContainer>
-    <!-- <content here> -->
-    {{ props.data }}
-  </VisualizationContainer>
+  <!-- <content here> -->
+  {{ props.data }}
 </template>
 
 <style scoped>
